requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
ctranslate2==4.6.0
deep-translator==1.11.4
deepl==1.23.0
fastapi==0.120.0
faster-whisper==1.2.0
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-translate==3.22.0
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.3.4
onnxruntime==1.23.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yt-dlp==2025.10.22
